+++ pos_pr/__manifest__.py
@@ -32,7 +32,6 @@
 
         'views/assets.xml',
         'views/invoice_payment_views.xml',
-        'views/pos_config_views.xml',
 
         'views/pos_session_views.xml',
         'views/account_journal_views.xml',
