+++ pos_pr/models/pos_config.py
@@ -8,12 +8,6 @@ class PosConfig(models.Model):
     _inherit = "pos.config"
 
     is_pos_pr_discount = fields.Boolean(default=False)
-    school_code_ids = fields.Many2many('school_base.school_code')
-    available_school_code_ids = fields.Many2many('school_base.school_code', compute='_compute_available_school_code_ids')
-
-    def _compute_available_school_code_ids(self):
-        for session in self:
-            session.available_school_code_ids = session.mapped('company_id.district_code_ids.school_code_ids')
 
     @api.constrains('payment_method_ids')
     def check_if_there_is_discount_payment_method(self):
